refactor: consolidate encryption logic into base encryption package
Quite a bit of logic was in encryption/api, which now only contains types.
